requirements.txt
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
neo4j==5.24.0
networkx==3.3
streamlit==1.39.0
langchain==0.3.1
langchain-community==0.3.1
python-dotenv==1.0.1
sendgrid==6.11.0
matplotlib==3.9.2
seaborn==0.13.2
